rearranged the packet structure within xdpprocessing and worked on the xdp processor

+++ src/uk/ac/cam/cl/juliet/slave/xdpprocessing/Packets/SymbolIndexMappingPacket.java
@@ -14,7 +14,7 @@ public class SymbolIndexMappingPacket extends Packet {
 	private long priceResolution;
 	private char roundLot;
 
-	public SymbolIndexMappingPacket(int[] data) {
+	public SymbolIndexMappingPacket(byte[] data) {
 		super(data);
 
 		this.symbolIndex = readLong(4);
